Lecture #16 (Path Order Matters)
In this lecture we learn that the order of the path matter
here if we make the "/posts/latest" below the "/posts/{id}" then we can never get the latest post.

to get the latest post, we have to code "/posts/latest"  this path above the "/posts/{id}" so that we can get the latest as well as unique post on the basis of "/posts/{id}"In FastAPI, the order of path matters.
If you place "/posts/{id}" above "/posts/latest", then "latest" will be treated as an id.
To fix this, always define specific paths like "/posts/latest" before dynamic ones like "/posts/{id}".